fix: PHPDoc improvements and Outcomes API restructure
- Added comprehensive class-level documentation to AbstractBaseApi, AbstractBaseDto, and HttpClient
- Fixed missing @package tags across all API classes (now match namespaces)
- Added missing @throws documentation where needed
- Fixed redundant type annotations in PHPDoc blocks
- Added complete documentation for str_to_snake_case() helper function
- Restructured Outcomes API for better organization:
  - Api/Outcomes/ for Outcome class
  - Api/OutcomeGroups/ for OutcomeGroup class
  - Api/OutcomeImports/ for OutcomeImport class
  - Api/OutcomeResults/ for OutcomeResult class
- Updated all namespaces, imports, and references to match new structure
- Fixed trailing whitespace issues from PHPDoc edits

All tests pass, PHPStan shows no errors, and PSR-12 coding standards are met.
