+++ tests/Api/Outcomes/OutcomeImportTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace CanvasLMS\Tests\Api\Outcomes;
 
-use CanvasLMS\Api\Outcomes\OutcomeImport\OutcomeImport;
+use CanvasLMS\Api\OutcomeImports\OutcomeImport;
 use CanvasLMS\Config;
 use CanvasLMS\Interfaces\HttpClientInterface;
 use PHPUnit\Framework\TestCase;
