+++ tests/Dto/Outcomes/Outcome/CreateOutcomeDTOTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace CanvasLMS\Tests\Dto\Outcomes\Outcome;
 
-use CanvasLMS\Dto\Outcomes\Outcome\CreateOutcomeDTO;
+use CanvasLMS\Dto\Outcomes\CreateOutcomeDTO;
 use PHPUnit\Framework\TestCase;
 
 class CreateOutcomeDTOTest extends TestCase
